requirements.txt
absl-py==1.2.0
anyio==3.6.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
docopt==0.6.2
fastapi==0.82.0
Flask==2.0.2
Flask-WTF==1.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.49.1
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.0.1
nltk==3.6.7
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
pip==21.2.2
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.1
pyparsing==3.0.9
python-dotenv==0.19.2
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==58.0.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
textblob==0.17.1
tqdm==4.62.3
typing_extensions==4.3.0
urllib3==1.26.8
uvicorn==0.18.3
Werkzeug==2.0.2
wheel==0.37.1
wincertstore==0.2
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
zipp==3.8.1
